requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pandas
requests
